web/docker_requirements.txt
# boto3
# celery
psycopg2-binary
channels
# daphne
django
django-cors-headers
django-import-export
djangorestframework
djoser
drf-yasg
gunicorn
pillow
python-decouple
requests
